Fix timing of GE and UPR counters
GE and UPR timers were off by one. Due to how they need to be calculated
they needed their own separate calculation.

GE occurs on 2025, 2030 and 2035. respectively, 12, 7 and 2 years to 2037.
so one needs to add at least 4 to ensure during the year it will still be
accurate.

UPR is similar by on 2026, 2031, 2036 so offset is by 5 instead.